+++ src/MeasureMap/SessionHandlers/WarmupSessionHandler.cs
@@ -25,12 +25,15 @@ namespace MeasureMap
 
             var stopwatch = Stopwatch.StartNew();
 
+            settings.IsWarmup = true;
             var ctx = settings.OnStartPipeline();
 
             task.Run(ctx);
 
             stopwatch.Stop();
             settings.OnEndPipeline(ctx);
+
+            settings.IsWarmup = false;
 
             var result = base.Execute(task, settings);